services/notification-service/requirements.txt
setuptools
flask==2.2.2
Werkzeug==2.2.2
flask-mail==0.9.1
python-dotenv==0.21.0
requests==2.28.1
numpy
